+++ src/Frontend/AddressbookManager.php
@@ -31,7 +31,7 @@ use MStilkerich\CardDavAddressbook4Roundcube\{Addressbook, Config};
 use MStilkerich\CardDavAddressbook4Roundcube\Db\AbstractDatabase;
 
 /**
- * @psalm-type ConfigurablePresetAttribute = 'name'|'url'|'username'|'password'|'active'|'refresh_time'
+ * @psalm-type ConfigurablePresetAttribute = 'name'|'url'|'username'|'password'|'active'|'refresh_time'|'rediscover_time'
  * @psalm-type Preset = array{
  *     name: string,
  *     url: string,
@@ -40,26 +40,51 @@ use MStilkerich\CardDavAddressbook4Roundcube\Db\AbstractDatabase;
  *     active: bool,
  *     use_categories: bool,
  *     readonly: bool,
+ *     rediscover_time: int,
  *     refresh_time: int,
  *     fixed: list<ConfigurablePresetAttribute>,
  *     require_always: list<string>,
- *     hide: bool,
- *     carddav_name_only: bool
+ *     hide: bool
  * }
- * @psalm-type AbookSettings = array{
+ * @psalm-type AccountSettings = array{
  *     name?: string,
  *     username?: string,
  *     password?: string,
  *     url?: string,
+ *     rediscover_time?: int,
+ *     active?: bool,
+ *     presetname?: string
+ * }
+ * @psalm-type AbookSettings = array{
+ *     name?: string,
+ *     url?: string,
+ *     account_id?: string,
+ *     active?: bool,
  *     refresh_time?: int,
  *     active?: bool,
  *     use_categories?: bool,
- *     presetname?: string
+ *     sync_token?: string,
+ *     discovered?: bool
  * }
+ * @psalm-import-type FullAccountRow from AbstractDatabase
  * @psalm-import-type FullAbookRow from AbstractDatabase
  */
 class AddressbookManager
 {
+    /**
+     * @var AccountSettings Template for account settings from the settings page.
+     *      The default values in this template also serve do determine the type (bool, int, string).
+     */
+    public const ACCOUNT_TEMPLATE = [
+        // standard account settings
+        'name' => '',
+        'username' => '',
+        'password' => '',
+        'url' => '',
+        'active' => true,
+        'rediscover_time' => 86400,
+    ];
+
     /**
      * @var AbookSettings Template for addressbook settings from the settings page.
      *      The default values in this template also serve do determine the type (bool, int, string).
@@ -68,26 +93,28 @@ class AddressbookManager
         // standard addressbook settings
         'name' => '',
         'url' => '',
-        'username' => '',
-        'password' => '',
         'active' => true,
         'use_categories' => true,
         'refresh_time' => 3600,
+        'discovered' => true,
     ];
 
     /**
-     * @var Preset Template for a preset; has the standard addressbook settings plus some extra properties.
+     * @var Preset Template for a preset; has the standard account/addressbook settings plus some extra properties.
      *      The default values in this template also serve do determine the type (bool, int, string, array).
      */
-    public const PRESET_TEMPLATE = AddressbookManager::ABOOK_TEMPLATE + [
+    public const PRESET_TEMPLATE = self::ACCOUNT_TEMPLATE + self::ABOOK_TEMPLATE + [
         // extra settings for presets
         'readonly' => false,
-        'carddav_name_only' => false,
         'hide' => false,
         'fixed' => [],
         'require_always' => [],
     ];
 
+    /** @var ?array<string, FullAccountRow> $accountsDb
+     *    Cache of the user's account DB entries. Associative array mapping account IDs to DB rows.
+     */
+    private $accountsDb = null;
 
     /** @var ?array<string, FullAbookRow> $abooksDb
      *    Cache of the user's addressbook DB entries. Associative array mapping addressbook IDs to DB rows.
@@ -105,6 +132,49 @@ class AddressbookManager
     {
         $infra = Config::inst();
         $admPrefs = $infra->admPrefs();
+
+        $config = $this->getAddressbookConfig($abookId);
+        $account = $this->getAccountConfig($config["account_id"]);
+        $presetname = $account["presetname"] ?? ""; // empty string is not a valid preset name
+
+        $readonly = !empty($admPrefs->presets[$presetname]["readonly"] ?? '0');
+        $requiredProps = $admPrefs->presets[$presetname]["require_always"] ?? [];
+
+        $account['username'] = Utils::replacePlaceholdersUsername($account["username"]);
+        $account['password'] = Utils::replacePlaceholdersPassword(Utils::decryptPassword($account["password"]));
+
+        $abook = new Addressbook($abookId, $account, $config, $readonly, $requiredProps);
+        return $abook;
+    }
+
+    /**
+     * Retrieves an account configuration (database row) by its database ID.
+     *
+     * @param string $accountId ID of the account
+     * @return FullAccountRow The addressbook config.
+     * @throws \Exception If no account with the given ID exists for this user.
+     */
+    public function getAccountConfig(string $accountId): array
+    {
+        $accounts = $this->getAccounts(false);
+
+        // check that this addressbook ID actually refers to one of the user's addressbooks
+        if (!isset($accounts[$accountId])) {
+            throw new \Exception("No carddav account with ID $accountId");
+        }
+
+        return $accounts[$accountId];
+    }
+
+    /**
+     * Retrieves an addressbook configuration (database row) by its database ID.
+     *
+     * @param string $abookId ID of the addressbook
+     * @return FullAbookRow The addressbook config.
+     * @throws \Exception If no addressbook with the given ID exists for this user.
+     */
+    public function getAddressbookConfig(string $abookId): array
+    {
         $abooks = $this->getAddressbooks(false);
 
         // check that this addressbook ID actually refers to one of the user's addressbooks
@@ -112,17 +182,47 @@ class AddressbookManager
             throw new \Exception("No carddav addressbook with ID $abookId");
         }
 
-        $config = $abooks[$abookId];
-        $presetname = $config["presetname"] ?? ""; // empty string is not a valid preset name
+        return $abooks[$abookId];
+    }
 
-        $readonly = !empty($admPrefs->presets[$presetname]["readonly"] ?? '0');
-        $requiredProps = $admPrefs->presets[$presetname]["require_always"] ?? [];
+    /**
+     * Updates some settings of an account in the database.
+     *
+     * If the given account ID does not refer to an account of the logged in user, nothing is changed.
+     *
+     * @param string $accountId ID of the account
+     * @param AccountSettings $pa Array with the settings to update
+     */
+    public function updateAccount(string $accountId, array $pa): void
+    {
+        // encrypt the password before storing it
+        if (isset($pa['password'])) {
+            $pa['password'] = Utils::encryptPassword($pa['password']);
+        }
 
-        $config['username'] = Utils::replacePlaceholdersUsername($config["username"]);
-        $config['password'] = Utils::replacePlaceholdersPassword(Utils::decryptPassword($config["password"]));
+        // optional fields
+        $qf = [];
+        $qv = [];
 
-        $abook = new Addressbook($abookId, $config, $readonly, $requiredProps);
-        return $abook;
+        foreach (array_keys(self::ACCOUNT_TEMPLATE) as $f) {
+            if (isset($pa[$f])) {
+                $v = $pa[$f];
+
+                $qf[] = $f;
+                if (is_bool($v)) {
+                    $qv[] = $v ? '1' : '0';
+                } else {
+                    $qv[] = (string) $v;
+                }
+            }
+        }
+
+        $userId = (string) $_SESSION['user_id'];
+        if (!empty($qf) && !empty($userId)) {
+            $db = Config::inst()->db();
+            $db->update(['id' => $accountId, 'user_id' => $userId], $qf, $qv, "accounts");
+            $this->accountsDb = null;
+        }
     }
 
     /**
@@ -135,11 +235,6 @@ class AddressbookManager
      */
     public function updateAddressbook(string $abookId, array $pa): void
     {
-        // encrypt the password before storing it
-        if (isset($pa['password'])) {
-            $pa['password'] = Utils::encryptPassword($pa['password']);
-        }
-
         // optional fields
         $qf = [];
         $qv = [];
@@ -157,19 +252,19 @@ class AddressbookManager
             }
         }
 
-        $userId = (string) $_SESSION['user_id'];
-        if (!empty($qf) && !empty($userId)) {
+        $accountIds = array_column($this->getAccounts(false), 'id');
+        if (!empty($qf) && !empty($accountIds)) {
             $db = Config::inst()->db();
-            $db->update(['id' => $abookId, 'user_id' => $userId], $qf, $qv, "addressbooks");
+            $db->update(['id' => $abookId, 'account_id' => $accountIds], $qf, $qv, "addressbooks");
             $this->abooksDb = null;
         }
     }
 
     /**
-     * Deletes the given addressbook from the database.
-     * @param string $abookId ID of the addressbook
+     * Deletes the given account from the database.
+     * @param string $accountId ID of the account
      */
-    public function deleteAddressbook(string $abookId): void
+    public function deleteAccount(string $accountId): void
     {
         $infra = Config::inst();
         $logger = $infra->logger();
@@ -177,6 +272,54 @@ class AddressbookManager
 
         try {
             $db->startTransaction(false);
+
+            // getAccountConfig() throws an exception if the ID is invalid / no account of the current user
+            $this->getAccountConfig($accountId);
+
+            $abookIds = array_column(
+                array_filter(
+                    $this->getAddressbooks(false),
+                    function (array $v) use ($accountId): bool {
+                        return $v["account_id"] == $accountId;
+                    }
+                ),
+                'id'
+            );
+
+            // we explicitly delete all data belonging to the account, since
+            // cascaded deletes are not supported by all database backends
+            foreach ($abookIds as $abookId) {
+                $this->deleteAddressbook($abookId);
+            }
+
+            $db->delete($accountId, 'accounts');
+
+            $db->endTransaction();
+        } catch (\Exception $e) {
+            $logger->error("Could not delete account " . $e->getMessage());
+            $db->rollbackTransaction();
+            throw $e;
+        } finally {
+            $this->abooksDb = null;
+        }
+    }
+
+    /**
+     * Deletes the given addressbook from the database.
+     * @param string $abookId ID of the addressbook
+     * @param bool   $skipTransaction If true, perform the operations without starting a transaction. Useful if the
+     *                                operation is called as part of an enclosing transaction.
+     */
+    public function deleteAddressbook(string $abookId, bool $skipTransaction = false): void
+    {
+        $infra = Config::inst();
+        $logger = $infra->logger();
+        $db = $infra->db();
+
+        try {
+            if (!$skipTransaction) {
+                $db->startTransaction(false);
+            }
 
             // we explicitly delete all data belonging to the addressbook, since
             // cascaded deletes are not supported by all database backends
@@ -197,20 +340,26 @@ class AddressbookManager
 
             $db->delete($abookId, 'addressbooks');
 
-            $db->endTransaction();
+            if (!$skipTransaction) {
+                $db->endTransaction();
+            }
         } catch (\Exception $e) {
             $logger->error("Could not delete addressbook: " . $e->getMessage());
-            $db->rollbackTransaction();
+            if (!$skipTransaction) {
+                $db->rollbackTransaction();
+            }
+            throw $e;
+        } finally {
+            $this->abooksDb = null;
         }
-
-        $this->abooksDb = null;
     }
 
     /**
-     * Inserts a new addressbook into the database.
-     * @param AbookSettings $pa Array with the settings for the new addressbook
+     * Inserts a new account into the database.
+     * @param AccountSettings $pa Array with the settings for the new account
+     * @return string Database ID of the newly created account
      */
-    public function insertAddressbook(array $pa): void
+    public function insertAccount(array $pa): string
     {
         $db = Config::inst()->db();
 
@@ -218,12 +367,10 @@ class AddressbookManager
         if (isset($pa['password'])) {
             $pa['password'] = Utils::encryptPassword($pa['password']);
         }
-
         $pa['user_id'] = (string) $_SESSION['user_id'];
-        $pa['sync_token'] = '';
 
         // required fields
-        $qf = ['name','username','password','url','user_id','sync_token'];
+        $qf = ['name','username','password','url','user_id'];
         $qv = [];
         foreach ($qf as $f) {
             if (!isset($pa[$f])) {
@@ -238,7 +385,7 @@ class AddressbookManager
         }
 
         // optional fields
-        $qfo = ['active','presetname','use_categories','refresh_time'];
+        $qfo = ['active','presetname','rediscover_time'];
         foreach ($qfo as $f) {
             if (isset($pa[$f])) {
                 $qf[] = $f;
@@ -252,8 +399,53 @@ class AddressbookManager
             }
         }
 
-        $db->insert("addressbooks", $qf, [$qv]);
+        $accountId = $db->insert("accounts", $qf, [$qv]);
+        $this->accountsDb = null;
+        return $accountId;
+    }
+
+    /**
+     * Inserts a new addressbook into the database.
+     * @param AbookSettings $pa Array with the settings for the new addressbook
+     * @return string Database ID of the newly created addressbook
+     */
+    public function insertAddressbook(array $pa): string
+    {
+        $db = Config::inst()->db();
+
+        // required fields
+        $qf = ['name','url','account_id'];
+        $qv = [];
+        foreach ($qf as $f) {
+            if (!isset($pa[$f])) {
+                throw new \Exception("Required parameter $f not provided for new addressbook");
+            }
+            $v = $pa[$f];
+            if (is_bool($v)) {
+                $qv[] = $v ? '1' : '0';
+            } else {
+                $qv[] = (string) $pa[$f];
+            }
+        }
+
+        // optional fields
+        $qfo = ['active','use_categories','refresh_time','sync_token','discovered'];
+        foreach ($qfo as $f) {
+            if (isset($pa[$f])) {
+                $qf[] = $f;
+
+                $v = $pa[$f];
+                if (is_bool($v)) {
+                    $qv[] = $v ? '1' : '0';
+                } else {
+                    $qv[] = (string) $pa[$f];
+                }
+            }
+        }
+
+        $abookId = $db->insert("addressbooks", $qf, [$qv]);
         $this->abooksDb = null;
+        return $abookId;
     }
 
     /**
@@ -294,6 +486,7 @@ class AddressbookManager
         // If the discovery URI points to an addressbook that is not part of the discovered ones, we only use that
         // addressbook
         try {
+            // FIXME replace this with an option for admin/user to manually add additional addressbooks to an account
             $directAbook = $infra->makeWebDavResource($uri, $account);
             if ($directAbook instanceof AddressbookCollection) {
                 $logger->debug("Only adding non-individual addressbook $uri");
@@ -308,25 +501,25 @@ class AddressbookManager
     }
 
     /**
-     * Returns all the user's addressbooks, optionally filtered.
+     * Returns all the user's accounts, optionally filtered.
      *
-     * @param $activeOnly If true, only the active addressbooks of the user are returned.
-     * @param $presetsOnly If true, only the addressbooks created from an admin preset are returned.
-     * @return array<string, FullAbookRow>
+     * @param $activeOnly If true, only the active accounts of the user are returned.
+     * @param $presetsOnly If true, only the accounts created from an admin preset are returned.
+     * @return array<string, FullAccountRow>
      */
-    public function getAddressbooks(bool $activeOnly = true, bool $presetsOnly = false): array
+    public function getAccounts(bool $activeOnly = true, bool $presetsOnly = false): array
     {
-        if (!isset($this->abooksDb)) {
-            $db = Config::inst()->db();
+        $db = Config::inst()->db();
 
-            $this->abooksDb = [];
-            /** @var FullAbookRow $abookrow */
-            foreach ($db->get(['user_id' => (string) $_SESSION['user_id']], [], 'addressbooks') as $abookrow) {
-                $this->abooksDb[$abookrow["id"]] = $abookrow;
+        if (!isset($this->accountsDb)) {
+            $this->accountsDb = [];
+            /** @var FullAccountRow $accrow */
+            foreach ($db->get(['user_id' => (string) $_SESSION['user_id']], [], 'accounts') as $accrow) {
+                $this->accountsDb[$accrow["id"]] = $accrow;
             }
         }
 
-        $result = $this->abooksDb;
+        $result = $this->accountsDb;
 
         if ($activeOnly) {
             $result = array_filter($result, function (array $v): bool {
@@ -337,6 +530,41 @@ class AddressbookManager
         if ($presetsOnly) {
             $result = array_filter($result, function (array $v): bool {
                 return !empty($v["presetname"]);
+            });
+        }
+
+        return $result;
+    }
+
+    /**
+     * Returns all the user's addressbooks, optionally filtered.
+     *
+     * @param $activeOnly If true, only the active addressbooks of the user are returned.
+     * @param $presetsOnly If true, only the addressbooks created from an admin preset are returned.
+     * @return array<string, FullAbookRow>
+     */
+    public function getAddressbooks(bool $activeOnly = true, bool $presetsOnly = false): array
+    {
+        $db = Config::inst()->db();
+
+        $allAccountIds = array_keys($this->getAccounts(false));
+        $accounts = $this->getAccounts($activeOnly, $presetsOnly);
+
+        if (!isset($this->abooksDb)) {
+            $this->abooksDb = [];
+            /** @var FullAbookRow $abookrow */
+            foreach ($db->get(['account_id' => $allAccountIds], [], 'addressbooks') as $abookrow) {
+                $this->abooksDb[$abookrow["id"]] = $abookrow;
+            }
+        }
+
+        $result = array_filter($this->abooksDb, function (array $v) use ($accounts): bool {
+            return isset($accounts[$v["account_id"]]);
+        });
+
+        if ($activeOnly) {
+            $result = array_filter($result, function (array $v): bool {
+                return $v["active"] == "1";
             });
         }